+++ components/server/src/database/datamodels.py
@@ -10,7 +10,7 @@ from server_utilities.functions import iso_timestamp
 
 def latest_datamodel(database: Database, max_iso_timestamp: str = ""):
     """Return the latest data model."""
-    timestamp_filter = dict(timestamp={"$lte": max_iso_timestamp or iso_timestamp()})
+    timestamp_filter = dict(timestamp={"$lte": max_iso_timestamp}) if max_iso_timestamp else None
     if data_model := database.datamodels.find_one(timestamp_filter, sort=[("timestamp", pymongo.DESCENDING)]):
         data_model["_id"] = str(data_model["_id"])
     return data_model or {}