+++ components/server/src/server_utilities/functions.py
@@ -23,7 +23,7 @@ def iso_timestamp() -> str:
 def report_date_time() -> str:
     """Return the report date requested as query parameter."""
     report_date_string = dict(bottle.request.query).get("report_date")
-    return str(report_date_string).replace("Z", "+00:00") if report_date_string else iso_timestamp()
+    return str(report_date_string).replace("Z", "+00:00") if report_date_string else ""
 
 
 def uuid() -> ReportId: